src/requirements.txt
fastapi
sqlalchemy 
python-dotenv 
uvicorn 
psycopg2-binary
gunicorn 
python-multipart 
pydantic[all] 
pydantic==1.10.8
email-validator 
passlib 
bcrypt
pydub==0.25.1
requests
starlette==0.19.1
starlette-context==0.3.5
python-jose==3.3.0
boto3
numpy==1.24.3
elasticsearch==7.14.0
aiohttp==3.10.9
faker==30.8.0
googlemaps==4.10.0